refactor: extract set and delete doc assertions
