+++ src/libs/satellite/src/db/mod.rs
@@ -1,3 +1,4 @@
+mod assert;
 pub mod impls;
 mod msg;
 mod state;
